Remove unused fields in pos transactions and set chain id.

+++ core/src/pos/types/src/account_config/resources/chain_id.rs
@@ -11,7 +11,7 @@ use serde::Deserialize;
 
 #[derive(Deserialize)]
 pub struct ChainIdResource {
-    chain_id: u8,
+    chain_id: u64,
 }
 
 impl ChainIdResource {
